feat(st-clk): move gate_refcounts check to use UINT8_MAX
The gate_refcounts check in the _clk_stm32_enable_core function
was using UINT_MAX, which is incorrect. It should use UINT8_MAX instead.
This change fixes the check to use the correct maximum value.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/393870
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>

+++ drivers/st/clk/clk-stm32-core.c
@@ -472,7 +472,7 @@ static int _clk_stm32_enable_core(struct stm32_clk_priv *priv, int id)
 
 	priv->gate_refcounts[id]++;
 
-	if (priv->gate_refcounts[id] == UINT_MAX) {
+	if (priv->gate_refcounts[id] == UINT8_MAX) {
 		ERROR("%s: %d max enable count !", __func__, id);
 		panic();
 	}
